pyproject.toml
[project]
name = "NullReceiver-Server"
version = "0.2.0"
description = "Server app for store minimum data for identify user(nickname/hash_password) and implement functionality for P2P connection between NullReceiver-Clients"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "alembic>=1.16.2",
    "asyncpg>=0.30.0",
    "bcrypt>=4.3.0",
    "fastapi>=0.115.13",
    "psycopg2>=2.9.10",
    "pydantic-settings>=2.10.1",
    "pyjwt[crypto]>=2.10.1",
    "redis>=6.2.0",
    "sqlalchemy>=2.0.41",
    "uvicorn>=0.34.3",
    "websockets>=15.0.1",
]
